Intro OCR(Optical Character Recognition)
OCR (Optical Character Recognition), is like a super-smart robot that can read words on paper or on a screen, just like a person does. It turns those words into text that computers can understand. So, if you take a picture of a page from a book, OCR will convert the words in the picture into a text document that you can edit, search, or share.